+++ noodles-sam/src/alignment/record_buf/name.rs
@@ -1,38 +1,40 @@
+use bstr::{BStr, BString};
+
 /// An alignment record name buffer.
 #[derive(Clone, Debug, Eq, Hash, PartialEq)]
-pub struct Name(Vec<u8>);
+pub struct Name(BString);
 
-impl AsRef<[u8]> for Name {
-    fn as_ref(&self) -> &[u8] {
-        &self.0
+impl AsRef<BStr> for Name {
+    fn as_ref(&self) -> &BStr {
+        self.0.as_ref()
     }
 }
 
-impl AsMut<Vec<u8>> for Name {
-    fn as_mut(&mut self) -> &mut Vec<u8> {
+impl AsMut<BString> for Name {
+    fn as_mut(&mut self) -> &mut BString {
         &mut self.0
     }
 }
 
 impl From<&[u8]> for Name {
     fn from(buf: &[u8]) -> Self {
-        Self::from(Vec::from(buf))
+        Self::from(BString::from(buf))
     }
 }
 
 impl<const N: usize> From<&[u8; N]> for Name {
     fn from(buf: &[u8; N]) -> Self {
-        Self::from(buf.as_slice())
+        Self::from(BString::from(buf.as_slice()))
     }
 }
 
-impl From<Vec<u8>> for Name {
-    fn from(buf: Vec<u8>) -> Self {
+impl From<BString> for Name {
+    fn from(buf: BString) -> Self {
         Self(buf)
     }
 }
 
-impl From<Name> for Vec<u8> {
+impl From<Name> for BString {
     fn from(name: Name) -> Self {
         name.0
     }